requirements.txt
Django==4.2.1
djangorestframework==3.14.0
sqlparse==0.4.4
psycopg2-binary==2.9.6
asgiref==3.6.0
pytz==2023.3
sqlparse==0.4.4